requirements.txt
aiobotocore==2.13.0
aiofiles==23.2.1
aiohttp==3.9.5
aioitertools==0.11.0
aioredis==2.0.1
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
astor==0.8.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bandit==1.7.9
billiard==4.2.0
botocore==1.34.106
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.5.3
Cython==3.0.10
darglint==1.8.1
dnspython==2.6.1
docutils==0.21.2
email_validator==2.1.1
eradicate==2.3.0
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cache2==0.2.1
fastapi-cli==0.0.3
flower==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
kombu==5.3.7
loguru==0.7.2
Mako==1.3.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
orjson==3.10.3
packaging==24.0
pathspec==0.12.1
pbr==6.0.0
pendulum==3.0.0
pep8-naming==0.13.3
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
pycodestyle==2.12.0
pydantic==2.7.1
pydantic-extra-types==2.7.0
pydantic-settings==2.2.1
pydantic_core==2.18.2
pydentic==0.0.1.dev3
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
python-stdnum==1.20
pytz==2024.1
PyYAML==6.0.1
redis==4.6.0
requests==2.32.3
restructuredtext_lint==1.4.0
rich==13.7.1
sentry-sdk==2.5.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
SQLAlchemy==2.0.30
starlette==0.37.2
starlette-csrf==3.0.0
stevedore==5.2.0
time-machine==2.14.1
tomli==2.0.1
tornado==6.4.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
wrapt==1.16.0
yarl==1.9.4